backend/requirements.txt
# FastAPI core
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2

# Data & analytics
numpy>=1.26
pandas>=2.2
scipy>=1.11
scikit-learn>=1.4
yfinance>=0.2.40
yahooquery>=2.3
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
pytz>=2023.3
tqdm>=4.66

# Massive flat files (S3-compatible)
boto3>=1.34
python-dotenv>=1.0

# Earnings scanner
selenium>=4.15
webdriver-manager>=4.0

# Background jobs
apscheduler>=3.10
